requirements.txt
Flask
slack_sdk
google-generativeai
